refactor: Community 관련 컴포넌트 Community 폴더로 이동
