+++ src/components/Community/CommunityItem/CommunityItem.styles.ts
