+++ src/components/Community/CommunityItem/CommunityItem.tsx
