--- src/components/CommunityDetail/CommunityDetail.tsx
@@ -1,9 +1,0 @@
-export const CommunityDetail = ({ category }) => {
-  return (
-    <div>
-      <div>
-        <h1></h1> <h2>{category}</h2>
-      </div>
-    </div>
-  );
-};
