+++ src/pages/menu/MenuList/MenuList.tsx
@@ -4,7 +4,7 @@ import { MenuListContext } from 'contexts/menuListContext';
 import { useParams } from 'react-router-dom';
 
 import { CertificatePageList } from 'components/Certificate/CertificatePageList/CertificatePageList';
-import { CommunityPageList } from 'components/CommunityItem/CommunityPageList';
+import { CommunityPageList } from 'components/Community/CommunityItem/CommunityPageList';
 import { InternPageList } from 'components/InternPageList/InternPageList';
 import { LanguagePageList } from 'components/LanguagePageList/LanguagePageList';
